Chore: service.py 파일 분리#17
