+++ review/review_service.py
@@ -1,43 +1,9 @@
+from .models import *
 from .response import *
-from .serializers import *
-from .s3_manager import *
+from .serializers import ReviewSerializer
+from .image_service import *
 
 
-### 이미지 관련 메소드
-def save_images_db(images, rid):
-    s3_urls = []
-    for image in images:
-        image_serializer = ImageSerializer(data={"image": image, "review":rid})
-        if not image_serializer.is_valid():
-            print("이미지 형식이 아님")
-            return None
-        
-        s3_url = save_image_s3(image, rid)
-        if s3_url is None:
-            return None
-        
-        s3_urls.append(s3_url)
-        image_serializer.validated_data["image"] = s3_url
-        image_serializer.save()
-    return s3_urls
-
-def delete_images_db(rid):
-    try:
-        review = Review.objects.get(rid=rid)
-        images = Image.objects.filter(review=review)
-        delete_image_s3(images)
-        images.delete()
-        return 204
-    except Review.DoesNotExist:
-        return 404
-    except Exception as e:
-        print("기존 이미지 삭제 에러 발생:", e)
-        return 500  # 삭제 실패 또는 예외 발생
-
-
-
-### 리뷰 관련 메소드
-# 리뷰 생성
 def post_review(request) -> ResponseDto:
     images = request.data.getlist('images', [])  # 기본 값을 빈 리스트로 지정
     if len(images[0]) == 0:     # 자동으로 ['']이 들어가기 때문에 한번 더 체크
@@ -63,7 +29,6 @@ def post_review(request) -> ResponseDto:
     return ResponseDto(status=201, data=data, msg=message['ReviewCreateSuccess'])
 
 
-# 리뷰 수정
 def put_review(request, review) -> ResponseDto:
     images = request.data.getlist('images', [])  # 기본 값을 빈 리스트로 지정
     if len(images[0]) == 0:     # 자동으로 ['']이 들어가기 때문에 한번 더 체크
@@ -90,7 +55,6 @@ def put_review(request, review) -> ResponseDto:
     return ResponseDto(status=200, data=data, msg=message['ReviewPutSuccess'])
 
 
-# 리뷰 리스트 조회
 def get_review_list() -> ResponseDto:
     review_list = []
     reviews = Review.objects.all()
@@ -102,7 +66,6 @@ def get_review_list() -> ResponseDto:
     return ResponseDto(status=200, data=review_list, msg=message['ReviewListGetSuccess'])
 
 
-# 단일 리뷰 조회
 def get_one_review(rid) -> ResponseDto:
     try:
         review = Review.objects.get(rid=rid)
@@ -114,7 +77,6 @@ def get_one_review(rid) -> ResponseDto:
         return ResponseDto(status=404, msg=message['ReviewNotFound'])
 
 
-# 리뷰 삭제
 def delete_review(rid):
     res = delete_images_db(rid)
     if res == 204:   # 삭제 성공
@@ -124,31 +86,3 @@ def delete_review(rid):
     else:
         return ResponseDto(status=500, msg=message['ReviewDeleteFailed'])
 
-
-
-### 댓글 관련 메소드
-def get_comment(rid) -> ResponseDto:
-    try:
-        review = Review.objects.get(rid=rid)
-        comments = Comment.objects.filter(review=review)
-        serializer = CommentSerializer(comments, many=True)
-        return ResponseDto(status=200, data=serializer.data, msg=message['CommentGetSuccess'])
-    except Review.DoesNotExist:
-        return ResponseDto(status=404, msg=message['ReviewNotFound'])
-
-def save_comment(request, rid, comment):
-    request.data['review'] = rid
-    request.data['writer'] = request.user.id    # 현재 로그인된 user
-    
-    if comment is None:     # post일 때
-        serializer = CommentSerializer(data=request.data)
-    else:   # put일 때
-        serializer = CommentSerializer(comment, data=request.data)
-    
-    if serializer.is_valid():
-        serializer.save()
-        if comment is None:
-            return ResponseDto(status=201, data=serializer.data, msg=message['CommentCreateSuccess'])
-        else:
-            return ResponseDto(status=200, data=serializer.data, msg=message['CommentPutSuccess'])
-    return ResponseDto(status=400, msg=message['InvalidFormat'])
